switch to eclipse-temurin and remove deprecated package objects
