+++ .scalafmt.conf
@@ -8,3 +8,4 @@ rewrite.rules = [
   SortModifiers
 ]
 maxColumn = 120
+docstrings.style = AsteriskSpace
