+++ core/src/test/scala/com/firstbird/emergence/core/utils/ConfigSpec.scala
@@ -1,7 +1,9 @@
 package com.fgrutsch.emergence.core.utils
 
 import cats.syntax.all.*
+import com.fgrutsch.emergence.core.utils.config.configFromYaml
 import com.fgrutsch.emergence.core.utils.config.given
+import com.typesafe.config.ConfigFactory
 import io.circe.{DecodingFailure, Json}
 import org.scalatest.prop.TableDrivenPropertyChecks
 import testutil.BaseSpec
@@ -10,7 +12,7 @@ import scala.concurrent.duration.*
 
 class ConfigSpec extends BaseSpec with TableDrivenPropertyChecks {
 
-  test("decode BuildStatusState successfully") {
+  test("decode FiniteDuration successfully") {
     val table = Table(
       "input"     -> "expected",
       "2 seconds" -> 2.seconds.asRight,
@@ -25,6 +27,55 @@ class ConfigSpec extends BaseSpec with TableDrivenPropertyChecks {
       val result    = jsonInput.as[FiniteDuration]
       result mustBe { expected }
     }
+  }
+
+  test("configFromYml converts yml to typesafe config") {
+    val yml = """
+    |repositories:
+    |   - name: owner/name
+    |     conditions:
+    |       - "target-branch == master"
+    |     merge:
+    |       strategy: merge-commit
+    |       close_source_branch: false
+    |
+    |defaults:
+    |   conditions:
+    |     - "build-success-all"
+    |     - "author == test"
+    |   merge:
+    |     strategy: squash
+    |     close_source_branch: true
+    """.stripMargin
+
+    val config = ConfigFactory.parseString("""
+    |repositories = [
+    |   {
+    |       "name" = "owner/name"
+    |       "conditions" = [
+    |           "target-branch == master"
+    |       ]
+    |       "merge" {
+    |           "strategy" = "merge-commit"
+    |           "close_source_branch" = false
+    |       }
+    |   }
+    |]
+    |
+    |defaults = {
+    |   conditions = [
+    |       "build-success-all",
+    |       "author == test"
+    |   ]
+    |   merge {
+    |       "strategy" = "squash"
+    |       "close_source_branch" = true
+    |   }
+    |}
+    """.stripMargin)
+
+    val result = configFromYaml(yml)
+    result.value mustBe { config }
   }
 
 }
